Adding Maven Wrapper support. (#50)
* Adding Maven Wrapper support.

This removes the need to have Maven installed locally and ensures
that the same Maven version is used by all developers.

To run Maven, replace `mvn` with `./mvnw`. For example:

    ./mvnw clean install

More information on https://github.com/takari/maven-wrapper

* Introduce discovery-parent-pom for standard build configuration.

This introduces some standard build configuration, including static
analysis. As new checks are performed as part of the build, some
cleanup was needed to existing code. The pom.xml was simplified
since some of the configuration is now available through the parent
pom.

For more details / documentation, see:
 https://github.com/wikimedia/wikimedia-discovery-discovery-parent-pom

Co-authored-by: Guillaume Lederrey <guillaume.lederrey@wikimedia.org>